Move actor api level out of internal.
